+++ pub/file/lib/chroot.dart
@@ -1,7 +1,6 @@
-/// A file system implementation that provides a view onto another file
-/// system, taking a path in the underlying file system, and making that the
-/// apparent root of the new file system. This is similar in concept to a
-/// `chroot` operation on Linux operating systems. Such a modified file system
-/// cannot name (and therefore normally cannot access) files outside the
-/// designated directory tree.
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+/// A file system that provides a view into _another_ [FileSystem] via a path.
 export 'src/backends/chroot.dart';