+++ pub/file/lib/src/backends/chroot/chroot_directory.dart
@@ -1,7 +1,13 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 part of file.src.backends.chroot;
 
 class _ChrootDirectory extends _ChrootFileSystemEntity<Directory, io.Directory>
     with ForwardingDirectory {
+  _ChrootDirectory(ChrootFileSystem fs, String path) : super(fs, path);
+
   factory _ChrootDirectory.wrapped(
     ChrootFileSystem fs,
     Directory delegate, {
@@ -10,8 +16,6 @@ class _ChrootDirectory extends _ChrootFileSystemEntity<Directory, io.Directory>
     String localPath = fs._local(delegate.path, relative: relative);
     return new _ChrootDirectory(fs, localPath);
   }
-
-  _ChrootDirectory(ChrootFileSystem fs, String path) : super(fs, path);
 
   @override
   FileSystemEntityType get expectedType => FileSystemEntityType.DIRECTORY;