+++ pub/file/lib/src/backends/chroot/chroot_file_system_entity.dart
@@ -1,3 +1,7 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 part of file.src.backends.chroot;
 
 abstract class _ChrootFileSystemEntity<T extends FileSystemEntity,
@@ -9,6 +13,12 @@ abstract class _ChrootFileSystemEntity<T extends FileSystemEntity,
   final String path;
 
   _ChrootFileSystemEntity(this.fileSystem, this.path);
+
+  @override
+  String get dirname => fileSystem.path.dirname(path);
+
+  @override
+  String get basename => fileSystem.path.basename(path);
 
   @override
   D get delegate => getDelegate();
@@ -66,7 +76,7 @@ abstract class _ChrootFileSystemEntity<T extends FileSystemEntity,
 
   @override
   String resolveSymbolicLinksSync() =>
-      fileSystem._resolve(path, notFound: _NotFoundBehavior.THROW);
+      fileSystem._resolve(path, notFound: _NotFoundBehavior.throwError);
 
   @override
   Future<FileStat> stat() {
@@ -93,7 +103,7 @@ abstract class _ChrootFileSystemEntity<T extends FileSystemEntity,
   @override
   Future<T> delete({bool recursive: false}) async {
     String path = fileSystem._resolve(this.path,
-        followLinks: false, notFound: _NotFoundBehavior.THROW);
+        followLinks: false, notFound: _NotFoundBehavior.throwError);
 
     String real(String path) => fileSystem._real(path, resolve: false);
     Future<FileSystemEntityType> type(String path) =>
@@ -104,7 +114,7 @@ abstract class _ChrootFileSystemEntity<T extends FileSystemEntity,
         await fileSystem.delegate.link(real(path)).delete();
       } else {
         String resolvedPath = fileSystem._resolve(p.basename(path),
-            from: p.dirname(path), notFound: _NotFoundBehavior.ALLOW_AT_TAIL);
+            from: p.dirname(path), notFound: _NotFoundBehavior.allowAtTail);
         if (!recursive && await type(resolvedPath) != expectedType) {
           String msg = expectedType == FileSystemEntityType.FILE
               ? 'Is a directory'
@@ -123,7 +133,7 @@ abstract class _ChrootFileSystemEntity<T extends FileSystemEntity,
   @override
   void deleteSync({bool recursive: false}) {
     String path = fileSystem._resolve(this.path,
-        followLinks: false, notFound: _NotFoundBehavior.THROW);
+        followLinks: false, notFound: _NotFoundBehavior.throwError);
 
     String real(String path) => fileSystem._real(path, resolve: false);
     FileSystemEntityType type(String path) =>
@@ -134,7 +144,7 @@ abstract class _ChrootFileSystemEntity<T extends FileSystemEntity,
         fileSystem.delegate.link(real(path)).deleteSync();
       } else {
         String resolvedPath = fileSystem._resolve(p.basename(path),
-            from: p.dirname(path), notFound: _NotFoundBehavior.ALLOW_AT_TAIL);
+            from: p.dirname(path), notFound: _NotFoundBehavior.allowAtTail);
         if (!recursive && type(resolvedPath) != expectedType) {
           String msg = expectedType == FileSystemEntityType.FILE
               ? 'Is a directory'