+++ pub/file/lib/src/backends/chroot/chroot_link.dart
@@ -1,7 +1,13 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 part of file.src.backends.chroot;
 
 class _ChrootLink extends _ChrootFileSystemEntity<Link, io.Link>
     with ForwardingLink {
+  _ChrootLink(ChrootFileSystem fs, String path) : super(fs, path);
+
   factory _ChrootLink.wrapped(
     ChrootFileSystem fs,
     io.Link delegate, {
@@ -10,8 +16,6 @@ class _ChrootLink extends _ChrootFileSystemEntity<Link, io.Link>
     String localPath = fs._local(delegate.path, relative: relative);
     return new _ChrootLink(fs, localPath);
   }
-
-  _ChrootLink(ChrootFileSystem fs, String path) : super(fs, path);
 
   @override
   Future<bool> exists() => delegate.exists();