+++ pub/file/lib/src/backends/local/local_file_system_entity.dart
@@ -1,3 +1,7 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 part of file.src.backends.local;
 
 abstract class _LocalFileSystemEntity<T extends FileSystemEntity,
@@ -9,6 +13,12 @@ abstract class _LocalFileSystemEntity<T extends FileSystemEntity,
   final D delegate;
 
   _LocalFileSystemEntity(this.fileSystem, this.delegate);
+
+  @override
+  String get dirname => fileSystem.path.dirname(path);
+
+  @override
+  String get basename => fileSystem.path.basename(path);
 
   @override
   Directory wrapDirectory(io.Directory delegate) =>