+++ pub/file/lib/src/backends/memory/memory_file_stat.dart
@@ -1,7 +1,12 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 part of file.src.backends.memory;
 
 class _MemoryFileStat implements io.FileStat {
-  static const _notFound = const _MemoryFileStat._internalNotFound();
+  static const _MemoryFileStat _notFound =
+      const _MemoryFileStat._internalNotFound();
 
   @override
   final DateTime changed;
@@ -41,7 +46,16 @@ class _MemoryFileStat implements io.FileStat {
   @override
   String modeString() {
     int permissions = mode & 0xFFF;
-    var codes = const ['---', '--x', '-w-', '-wx', 'r--', 'r-x', 'rw-', 'rwx'];
+    List<String> codes = const <String>[
+      '---',
+      '--x',
+      '-w-',
+      '-wx',
+      'r--',
+      'r-x',
+      'rw-',
+      'rwx',
+    ];
     List<String> result = <String>[];
     result
       ..add(codes[(permissions >> 6) & 0x7])