+++ pub/file/lib/src/backends/memory/memory_file_system.dart
@@ -1,10 +1,14 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 part of file.src.backends.memory;
 
 const String _separator = '/';
 const String _thisDir = '.';
 const String _parentDir = '..';
 
-/// Visitor callback for use with [_findNode].
+/// Visitor callback for use with `_findNode`.
 ///
 /// [parent] is the parent node of the current path segment and is guaranteed
 /// to be non-null.
@@ -18,10 +22,10 @@ const String _parentDir = '..';
 /// do things like recursively create or delete folders.
 ///
 /// [currentSegment] is the index of the current segment within the overall
-/// path that's being walked by [_findNode].
+/// path that's being walked by `_findNode`.
 ///
 /// [finalSegment] is the index of the final segment that will be walked by
-/// [_findNode].
+/// `_findNode`.
 typedef _Node _SegmentVisitor(
   _DirectoryNode parent,
   String childName,
@@ -47,21 +51,26 @@ class MemoryFileSystem extends FileSystem {
   String _systemTemp;
   String _cwd = _separator;
 
+  /// Creates a new `MemoryFileSystem`.
+  ///
+  /// The file system will be empty, and the current directory will be the
+  /// root directory.
   MemoryFileSystem() {
     _root = new _RootNode(this);
   }
 
   @override
-  Directory directory(path) => new _MemoryDirectory(this, common.getPath(path));
+  Directory directory(dynamic path) =>
+      new _MemoryDirectory(this, common.getPath(path));
 
   @override
-  File file(path) => new _MemoryFile(this, common.getPath(path));
+  File file(dynamic path) => new _MemoryFile(this, common.getPath(path));
 
   @override
-  Link link(path) => new _MemoryLink(this, common.getPath(path));
+  Link link(dynamic path) => new _MemoryLink(this, common.getPath(path));
 
   @override
-  String get pathSeparator => _separator;
+  p.Context get path => new p.Context(style: p.Style.posix, current: _cwd);
 
   /// Gets the system temp directory. This directory will be created on-demand
   /// in the root of the file system. Once created, its location is fixed for
@@ -142,9 +151,6 @@ class MemoryFileSystem extends FileSystem {
     }
     return node.type;
   }
-
-  /// Gets the path context for this file system given the current working dir.
-  p.Context get _context => new p.Context(style: p.Style.posix, current: _cwd);
 
   /// Gets the node backing for the current working directory. Note that this
   /// can return null if the directory has been deleted or moved from under our