+++ pub/file/lib/src/backends/memory/memory_link.dart
@@ -1,3 +1,7 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 part of file.src.backends.memory;
 
 class _MemoryLink extends _MemoryFileSystemEntity implements Link {
@@ -88,4 +92,7 @@ class _MemoryLink extends _MemoryFileSystemEntity implements Link {
 
   @override
   Link _clone(String path) => new _MemoryLink(fileSystem, path);
+
+  @override
+  String toString() => "MemoryLink: '$path'";
 }