+++ pub/file/lib/src/forwarding/forwarding_file_system_entity.dart
@@ -1,7 +1,13 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
 part of file.src.forwarding;
 
+/// A file system entity that forwards all methods and properties to a delegate.
 abstract class ForwardingFileSystemEntity<T extends FileSystemEntity,
     D extends io.FileSystemEntity> implements FileSystemEntity {
+  /// The entity to which this entity will forward all methods and properties.
   @protected
   D get delegate;
 
@@ -79,4 +85,10 @@ abstract class ForwardingFileSystemEntity<T extends FileSystemEntity,
 
   @override
   String get path => delegate.path;
+
+  @override
+  String get basename => fileSystem.path.basename(path);
+
+  @override
+  String get dirname => fileSystem.path.dirname(path);
 }