+++ pub/file/CHANGELOG.md
@@ -1,3 +1,18 @@
+#### 2.0.1
+
+* Minor doc updates
+
+#### 2.0.0
+
+* Improved `toString` implementations in file system entity classes
+* Added `ForwardingFileSystem` and associated forwarding classes to the
+  main `file` library
+* Removed `FileSystem.pathSeparator`, and added a more comprehensive
+  `FileSystem.path` property
+* Added `FileSystemEntity.basename` and `FileSystemEntity.dirname`
+* Added the `record_replay` library
+* Added the `testing` library
+
 #### 1.0.1
 
 * Added `FileSystem.systemTempDirectory`